flask-server/requirements.txt
Flask==2.1.0
go-openai==0.1.0
assemblyai==0.1.3